chore: refactor query builder
