+++ internal/sql/qb_delete.go
