+++ internal/sql/qb_insert.go
@@ -51,22 +51,3 @@ func (b *QueryBuilder) buildValues(table *schema.Table, rows []map[string]interf
 func (b *QueryBuilder) buildInsertInto(table *schema.Table) string {
 	return fmt.Sprintf("INSERT INTO %s (%s)", table.Name.Value, strings.Join(table.ColumnsNames(), ", "))
 }
-
-func (b *QueryBuilder) mapValue(val interface{}) string {
-	colValStr := "null"
-
-	switch tColVal := val.(type) {
-	case string:
-		colValStr = fmt.Sprintf("'%s'", tColVal)
-	case bool:
-		if tColVal {
-			colValStr = "true"
-		} else {
-			colValStr = "false"
-		}
-	case int, int8, int16, int32, int64, uint, uint8, uint32, uint64:
-		colValStr = fmt.Sprintf("%d", tColVal)
-	}
-
-	return colValStr
-}
