+++ internal/sql/qb_insert_test.go
